fix: re-queue disabled jobs on restart, #53
@olmaga

As per my understanding, if a job fails during execution, it is
[de-queued](https://github.com/ortoloco/ortoloco/blob/c0afa21042a543fe7f9785886dff187cf4304344/django_cron/base.py#L126)
and the queuing state is persisted in the DB.
When jobs are (re)registered (i.e. on system (re)start) they are restored
from the DB if they exist. In that case, the 'queued' flag must be reset to
True, or otherwise they will never be executed again.

Please validate and deploy the fix if you agree. This feature is
strongly missed by the garden team, since some people 'forget' to go to
the jobs they committed to.

Cheers.