Make generate_testdata_advanced work again
This management command was outdated and referred to models that
have been deleted in the meantime.

Also, it was using various real world datasources which have become
unavailable.

Update the command to work with current models and replace external
datasources with the faker library.

The generated data might not be as good, but will make this command
independend from the availability of external services.

Also, add a test so any breaking changes to this script will not go
unnoticed again.