travis: Use requirement-dev.txt to track all dev requirements
We recently added faker as a dependency that needs to be available
when we run tests.

Once we moved it from `requirements.txt` to `requirements-dev.txt`,
CI failed because faker was not being installed anymore.

So let's have Travis install all requirements from
`requirements-dev.txt`.

Not only does this make faker available for testing again, it also
allows us to move the installation of the other development requirements
(`coverage` and `django-coverage-plugin`) from the Travis config into
one consistent location.